# Faster SequenceRecord initialization.