@rhpvorderman 2022-10-29
The post above is quite a ramble. Let me summarize now that I have a more clear head:
Any python object needs to be allocated in memory. This is the purpose of the `__new__` function. The `__new__` function can also be used to initialize the object if its attributes are static.
The `__init__` function is used on objects on which its attributes are dynamic and context dependent. Python automatically calls the objects `__new__` function before `__init__` is called. 

Cython implements it as follows.
- The `__new__` function is called to initialize a cdef class. In this `__new__` function all python object members are set to None.
- The `__cinit__` function is called from the `__new__` function to initialize all the C members of the object (if the user has specified `__cinit__`). `__cinit__` does not take C arguments, but an a python argument tuple and keyword dict instead. Which are parsed.
- The `__init__` function is implemented by the user. This also takes a python argument tuple and keyword dict. 

By leaving out `__cinit__` entirely, it is not called anymore. That means that an extra step parsing an argument tuple and a keyword dict is left out. This is a major source of slowdown. We don't need it because there are no C attributes that need initializing in the SequenceRecord class.