@rhpvorderman 2022-11-02
>  and I think __init__ is indeed good enough.

By not calling `__cinit__` (via `__new__`), `__init__` should also be faster. Unfortunately this is hard to verify with timeit. The old `__cinit__` code was pages long, mostly for parsing arguments (it hardly did anything else). I am glad we have been able to optimize it out.