@rhpvorderman 2022-11-02
> Probably time for a release (0.9.2) afterwards. I’ll look into that.

I propose merging #96 too and bumping to 0.10.0. Then we can use that version for the benchmarks in the paper as it should be faster.